+++ common/views/components/CompactCard/CompactCard.tsx
@@ -5,7 +5,6 @@ import EventDateRange from '../EventDateRange/EventDateRange';
 import StatusIndicator from '../StatusIndicator/StatusIndicator';
 import LabelsList from '../LabelsList/LabelsList';
 import ImagePlaceholder from '../ImagePlaceholder/ImagePlaceholder';
-// $FlowFixMe(tsx)
 import ImageType from '../Image/Image';
 import { ColorSelection } from '../../../model/color-selections';
 import MediaObjectBase from '../MediaObjectBase/MediaObjectBase';
